Phase 2 commit 9: rename api.ts → server/index.ts; tighten domain layer
- server/api.ts → server/index.ts (composition root). The test
  file becomes server/index.test.ts. vite.config.ts updated.
- domain/yaml.ts no longer imports 'fs'. extractTitle (which read
  a file) becomes extractTitleFromContent (pure: takes the bytes
  + a fallback). The fs adapter wraps it in a tiny
  extractTitleFromFile helper that does the read.

This finishes the layer split: import-rule grep over
server/http and server/domain returns no fs / no fetch, and
http never imports a *.fs.ts adapter — only the composition
root in index.ts does.

Server tests (33) green.
